Make the game playable in the webpage
Three buttons are added to the html representing player's choices.
Click a button to play against the computer's choice. The result of each
round is displayed. When a player reaches 5 wins, the final result is
displayed. The next time any button is pressed, the results of the last
round are deleted except for the final result.